requirements.txt
requests
pandas
snowflake-connector-python
python-dotenv
